+++ src/locus.erl
@@ -631,11 +631,6 @@ parse_http_url(DatabaseURL) ->
     of
         false ->
             false;
-        {ok, {Scheme, "", "geolite.maxmind.com", Port,
-              "/download/geoip/database/GeoLite2-" ++ Suffix, _, _}}
-          when Scheme =:= http, Port =:= 80;
-               Scheme =:= https, Port =:= 443 ->
-            parse_discontinued_geolite2_http_url(DatabaseURL, Suffix, ByteList);
         {ok, _Result} ->
             {http, ByteList};
         {error, _Reason} ->
@@ -643,29 +638,6 @@ parse_http_url(DatabaseURL) ->
     catch
         error:badarg -> false
     end.
-
-parse_discontinued_geolite2_http_url(DatabaseURL, Suffix, ByteList) ->
-    case Suffix of
-        "Country.tar.gz" ->
-            log_warning_on_use_of_discontinued_geolite2_http_url(DatabaseURL, 'GeoLite2-Country'),
-            {maxmind, 'GeoLite2-Country'};
-        "City.tar.gz" ->
-            log_warning_on_use_of_discontinued_geolite2_http_url(DatabaseURL, 'GeoLite2-City'),
-            {maxmind, 'GeoLite2-City'};
-        "ASN.tar.gz" ->
-            log_warning_on_use_of_discontinued_geolite2_http_url(DatabaseURL, 'GeoLite2-ASN'),
-            {maxmind, 'GeoLite2-ASN'};
-        _ ->
-            {http, ByteList}
-    end.
-
-log_warning_on_use_of_discontinued_geolite2_http_url(LegacyURL, DatabaseEdition) ->
-    locus_logger:log_warning(
-      "Public access to GeoLite2 was discontinued on 2019-12-30"
-      "; converting legacy URL for your convenience.~n"
-      "Update your `:start_loader' and `:loader_child_spec' calls to silence this message.~n"
-      "(Use the tuple {maxmind, '~ts'} instead of the legacy URL \"~ts\")",
-      [DatabaseEdition, LegacyURL]).
 
 parse_filesystem_url(DatabaseURL) ->
     try unicode:characters_to_list(DatabaseURL) of